lionfish: avoid redundant loading calculations
On MSM8939, the load calculation routine would be performed multiple
times within a sampling period (once for each online CPU). What
happened was that the first load calculation within a sampling period
would be correct, while the others will be measured relative to a
point in time only a few microseconds before the last check. This
resulted in bogus load measurements of 0 and 100. This in turn led
to lionfish jumping to maximum frequency under mundane loading
conditions.

The new load calculation function will only calculate the load for a
CPU once per sampling period, and report the previous value if called
again within the same sampling period. This solves the incorrect load
calculation bug on MSM8939, and also improves the computational
efficiency of the governor.